Removed Connection dependency
Because we have decided not to manage health codes in the database, the
interface for the HealthCodeProvider should no longer exhibit a
dependency on an SQL Connection object.  The interface should not expose
how the underlying terms are managed, regardless of whether it uses
relational databases, a NoSQL DB or some collection of flat files.