Changed initialisation of health code providers
SQLHealthOutcomeManager has been changed to accept a RIFStartupOptions
object which may specify the file path of the web application it is
running in.  The manager advertises a number of example health code
providers, all of which are based on having them read terms from an XML
file.  The location of the files is therefore critical for the providers
to work correctly.  The health outcome manager now has in its contructor
an argument for RIFStartupOptions.  These options can specify the
location of the web application that the code is using.  Within that web
application folder is the location of the *.xml files which will provide
the data for the taxonomies.