Developed a Test service in the automated tests
Originally, the main service used in the test suites was the
ProductionRIFJobSubmissionService class.  The problem was that some
methods were useful for testing but should not be part of the service
API in production.  The solution was to develop a
TestRIFSubmissionService which could extend the
ProductionRIFJobSubmissionService and override protected test methods so
they could be called within test cases.