Methods to get min/max age group in investigation
An investigation has a collection of AgeBand objects, each of which has
a lower and upper AgeGroup.  Obtaining the minimum and maximum Age Group
from an Investigation is important for filling out form fields.  For
now, we have given only partial support for the idea of having multiple
AgeBands.  much of the database still operates with the assumption there
is only one.  So while the front end and middleware are able to assume
multiple age bands, only one age band is actually submitted -- the sum
of years covered by all the other AgeBand objects.