Changing sorter to hash on display name
DisplayableItemSorter sorts a collection of displayable items.  It makes
use of a HashMap that relies on the displayable item's getIdentifier()
value.  however, we are beginning to observe that the identifier field
in a lot of the domain classes tends to be null.  Therefore, we are
using a field which will not generate nulls - the value provided by a
displable item's getDisplayName() field.