Added method to get String array of area IDs
StudyArea and ComparisonArea objects manage collections of MapArea
objects.  However, when the map areas are specified for the database,
only the area identifier is needed (the label field is not).  As a
convenience method, I've put in getMapAreaIdentifiers() to support this
need.  It is used by the SQLRIFSubmissionManager class when a
RIFJobSubmission