Fixed addRIFJobSubmission() method, added clear...
Fixed the addRIFJobSubmission so that it would add parts of a
RIFJobSubmission to the underlying tables for studies, comparison areas,
study areas and investigations.  Yet to do is testing whether the
covariates and health codes work.  Also added a clear method to delete
all jobs in the database associated with a given user.