Added a new query formatter to call DB functions
In a growing number of API methods, the middleware is delegating to a
function that is part of the database schema.  This is a query formatter
used to help standardise the construction of SQL queries that call
functions.  The goal is to use formatter classes such as these to help
reduce the number of errors in the queries submitted to the database,
and to later add features which would help format the queries for
reports and displays.