Adding new methods to retrieve study results
The RIFJobSubmissionAPI has been extended to include more methods that
will be used to retrieve data from the tables that are generated when
the RIF processes a study.  Extending this class has meant extending the
main implementing class of RIFJobSubmissionAPI -
ProductionRIFJobSubmissionService.  Following the conventions of other
API methods, this service class delegates to a manager class which
actually constructs the SQL queries and sends them to the database.  To
serve this role,SQLResultsQueryManager was created.