Reworking service classes
Previously, the RIF relied on RIFJobSubmissionAPI, which ended up
growing to almost 50 methods as it began to include various methods that
were used for retrieving results.  The burden of implementing all API
methods fell to ProductionRIFJobSubmissionService, which began to become
too large to maintain.  As well, we had to consider protected methods
that could be overridden for testing purposes.  We now have the concept
of a "RIF study service bundle", which manages a collection of simpler
services, as well as tasks for logging in and out.  Code for the service
bundle is expressed by an AbstractRIFStudyServiceBundle class, and final
classes that are appropriate for production and testing environments.
The service bundle manages two services: one for study submission
(RIFStudySubmissionAPI) and another for retrieving study results
(RIFStudyResultRetrievalAPI).  Like the service bundle concept, these
services are supported by AbstractXXX, TestXXX and ProductionXXX service
classes which ensure that testing methods don't end up in production
code.