Reducing number of error codes
The RIF makes use of an application-level exception to support two
goals: (1) to provide intelligent, plain English error messages to
clients in a way that hides potentially sensitive data produced by the
database and (2) to provide an error code that makes it easy to test
exception cases in automated test suites.  Therefore, each
RIFServiceException has one or more error messages meant for people and
a RIFServiceError that is meant to help with comparing accepted and
actual error code return values in automated test suites.

Over the course of development, the number of different error codes has
grown significantly, and many of them describe the same category of
error: a query that fails to be executed by the database.  It has been
decided that while specific error messages are still useful to help
inform users of client applications, for the purposes of testing we only
need one kind of error code in RIFServiceError to report failings in the
database: the code "DATABASE_QUERY_FAILED"

I went through the manager classes and ensured that only this code was
used if a try-catch block caught an SQLException.