Encouraging policy of min visibility
Many classes have had the "public" modifier removed from the class
definition, so they are not visible outside the scope of their package.
This has been done throughout the rifServices and rifJobSubmissionTool
projects.  The change increases security and simplifies the collection
of classes that users of the code need care about.