Moving AgeGroupSortingOption enum from API
Previously, the AgeGroupSortingOption enum has been defined within the
definition of the APIs.  Both the RIFStudySubmissionAPI and the
RIFStudyResultRetrievalAPI have this enum defined so it makes sense to
extract the common enum.  I could have put the common definition into
the RIFStudyServiceAPI interface but I thought it was better to isolate
it as its own independent enumeration.