Changed "checkXXXExist(...)" checks
Changed the signatures of methods of the form "checkXXXExists" so that
they took an identifier field rather than a whole business class object.
This was done to help simplify the code of methods in the retrieval
services, where we only cared about the identifier for a given type of
object rather than having a full instance of one with all its other
fields.