Adding display name method
StudyResultRetrievalContext is really meant to be a simple data
container class to hold a set of identifier values that commonly appear
in the web service calls related to retrieving results.  When errors
occur it is often useful to have a display name for a parameter that was
involved in a call.  In this case, getDisplayName() does not produce a
name that is as easily readable as those of other business classes that
specifically relate to epidemiologist concepts (eg: health code, study,
covariate, etc.).  However, this class has been developed more to aid
the middleware more than provide a meaningful concept to people.  The
retrieval context object may need to be moved in future to another
package, given that a context here isn't really a business concept.