Major refactoring of code for services
Previously, all the methods used by both the study submission and study
retrieval class were defined in RIFJobSubmissionAPI and implemented by
the ProductionRIFJobSubmissionService class.  This lead to the bloat
occurring in both of these files.  Therefore, code for the services was
redistributed into two services: a RIFStudySubmission service and a
RIFStudyResultRetrieval service.   Methods common to the two services
were isolated into an abstract superclass. The classes used to wrap
results from the main Java APIs were themselves refactored in a fashion
similar to the way the code for the main services was done.  Much of the
code in the classes used to support result retrieval are heavily stubbed
and will be implemented through more focused changes later.