Proxy class for BoundaryRectangle
The sole purpose of this class is to package BoundaryRectangle data in a
way that is amenable to streaming as JSON data.  The class looks
slightly redundant alongside BoundaryRectangle but I've tried to make it
a policy to isolate all classes for web services in the
rifServices.restfulWebServices package.  I'm trying to keep the business
classes free of annotations or fields that would service a JSON output
that is needed by a particular type of front-end client.