Overhaul of automated tests
Initially, the test cases which exercised the service APIs were grouped
together by themes corresponding to business classes (eg:
"TestAgeGenderYearServiceFeatures".

Now, each API method has its own test suite.  The test cases are
generally designed to fail due to a minimum number of assert statements.
The sheer number of test cases which have been created warranted a
different approach to documentation

Now most of the meaning of a test case can be interpreted based on the
naming convention used for method names. Here are conventions:

[method name]_COMMON[1...n]: Common case test cases.  They use typical
data and obtain a typical result.

[method name]_UNCOMMON[1..n]: uncommon test case scenarios.  For
example, it would be unusual yet valid to fetch a block of calculated
results where the start and end indices are the same.

[method name]_NULL[1...n]: check for a null method parameter value.  n
will almost always correspond to the nth parameter in the method
[method name]_EMPTY[1...n]: checks whether a method parameter is (1) a
String with a value of the empty string "" or (2) a business class which
has a field that is blank or has another object with a blank field. n
refers to the nth parameter.

[method name]_NONEXISTENT[1...n]: uses a non-existent value for a
business object that is passed to the API method
[method name]_MALICIOUS[1...n]: attempts to detect a malicious String
field value.  The detection is recursive within a business object that
may in turn have other business objects.  n is the nth parameter.