Clarifying issue with GeoLevelView
We've had some dialogue about whether API methods need a GeoLevelView or
GeoLevelToMap reference.  GeoLevelView refers to the resolution that is
shown in a selectable display map.  The purpose of resolution here is
simply to make it easier to select areas on a map.  GeoLevelToMap refers
to the resolution that is expected in the result.  We've known this for
a long while, but it's been clarified more in some of the API methods
related to retrieving results.