Adding subproject for generic UI
The RIF Tool Suite will include at least three Swing-based desktop
applications: (1) the initial prototype of the RIF Study Submission Tool
(2) the RIF Data Loader Tool and (3) the RIF IT Governance Tool.  The
purpose of the rifGenericUILibrary is to consolidate convenience classes
that could be used to make the UIs of these three applications.  There
is minor dependence on the rifServices project, but at a later date this
may be removed.