Changing package for service APIs
I've had some trouble deciding whether the RIF service APIs belong in
the businessConceptLayer or the dataStorageLayer.  From the point of
view of isolating concepts that are not tied to storage, it makes sense
to put them there.  However, I've now moved them to the dataStorageLayer
so that the interfaces and not the service implementation classes can be
outside the package.