Modified classes for formatting SQL queries
I've added some new query formatter classes because the RIF Data Loader
Tool needed them and they were generic enough in nature that they might
be useful in other parts of the code base.  As well, query formatters
now have methods that allow query generation to be "pretty printed" with
comments and indentation levels.  The goal of these enhancements was to
enhance the readability of generated SQL code.