Adding web service methods
Both the RIFStudyResultRetrievalWebServiceResource and
RIFStudySubmissionWebServiceResource now both have methods that used to
be in AbstractRIFWebServiceResource.  However, it seems that Jersey web
services had difficulty accessing the methods in the abstract classes.
It now seems to work just fine now that each of them has annotations for
the same methods, which call non-annotated methods in
AbstractRIFWebServiceResource