Adding notion of types of security threats
Developing a way of further describing security threats so the
information can be used to make decisions about whether a user should be
blocked.  If it's a malicious attack, the user is blacklisted.  However,
if it's "suspicious behaviour", then we may want to think about allowing
N suspicious attempts before the person is blocked.  This measure helps
accommodate accidental access attempts vs ones that appear determined to
look for weaknesses in the web service APIs