Adding behaviour to logException(...)
When a security exception is logged, it is interrogated to determine
what kind of threat it is.  If it's a malicious code attack, then the
user is added to a black list so they cannot make further queries.  If
it's merely a suspicious action, then a counter is indicated for the
user.  If the counter exceeds a threshold, then the user will be added
to the black list.