Generalised RIFServiceException
When the RIFServiceException class was first created, it used two
attributes: error messages and an error code.  The error code relied on
the enumerated type RIFServiceError.  As other tools evolved, other
exception classes such as RIFDataLoaderToolException and
RIFJobSubmissionToolException evolved.  Each of these had so many
similarities with RIFServiceException that it led to repetitive code.

The key to streamlining code for these exception classes is to make
RIFServiceException manage an error code as an Object rather than as a
specific enumerated type that may vary from one RIF application to
another.