Added support for suspicious event counter
Adding support to distinguish between two types of security exceptions:
malicious code attacks and suspicious user patterns.  A suspicious user
pattern could be repeatedly trying to access a web resource when the
user is not logged in.  Whereas if a user did this once it could be
accidental, repeatedly doing it would suggest that someone or something
were trying to find weak points in the web service.  The class has been
modified so that it keeps a counter for the number of suspicious
attempts made by a user.  If the count exceeds a threshold, then the
user can be black listed.