Added geographicalIdentifier to MapArea
MapArea began with two properties: an identifier and a label.  Now we
have a third attribute called a geographicalIdentifier, which is a
geographicalIdentifier meant to be unique across all collections of map
areas used in the database.  Whereas identifier can have a variety of
characters in its value, geographicalIdentifier is meant to be an
integer and may well be converted to this type as the use of this field
matures in the system.