Ensure JSON result returns arrays
The web applications that use the middleware are written using
JavaScript.  It is convenient if all the web services return data as an
array, whether the array has one item or several.

As well, renaming references to mapAreaIdentifiers to
geographicalIdentifiers to make the variable more specific.  After all,
MapArea now has two fields: geographicalIdentifier and identifier.  We
have yet to determine how much these two fields will diverge but it is
clear that the field guaranteed to uniquely identify a map area will be
geographicalIdentifier, not identifier.
Modifying signature of getTiles method so it accepts a tile identifier
and a zoom factor.  These two attributes will be used to create a
caching system which helps reduce the work of retrieving tiles.