Allowing Close-only version
OKCloseButtonPanel almost always features two buttons: "OK", where
changes in form fields are committed to underlying data model objects
and "Close", where these changes are discarded.  In other cases, we are
simply presenting data rather than inviting the user to decide on saving
changes.  Therefore we only need a Close button.  We could implement
this simply as a "Close" JButton but perhaps we can capitalise on reuse
of other properties that would be set in the OKCloseButtonPanel