Eliminating reliance on parameterised UI method
Parts of the UI code use the notion of a number of columns or number of
rows to help instantiate JTextArea and JTextField components.  The
numbers used for these values seem arbitrary in which values are used
but are consistently applied throughout the UI code.  Therefore I
thought it was better to simply create default text fields and text
areas where the number of columns and rows were set by a default value
that could later be parameterised.