Supporting editing user ids
We feel that the tasks of creating, deleting or altering user IDs will
ultimately be a task managed by the IT Governance Tool when eventually
it is developed.  For now we need to have these features somewhere and
the DataLoaderTool is the next best place to support them.  Therefore
changes to the data loader service API, the creation of a "UserManager",
and the development of supporting UI classes have been carried out.