Making interfaces comply with name convention
Adopting the convention that if a code file describes an interface, its
name should end in Interface or API.  API will describe an interface
that publishes behaviour that is used by another application.  All other
interfaces will end in Interface.  The goal of this change is to promote
readability as the code base continues to grow and accumulate more
classes.