Modified signature of getCovariates(...) service call
Removed the parameter GeoLevelSelect from getCovariates(...) because we
don't need it.  We will assume for now that in any given geography, it
either has covariate data for a particular resolution across the country
or it has no covariate data for a particular resolution in any of the
country.  Note that this may change if, for example, a study comprising
two states has one which collects socio-economic level data at county
level and the other which does not collect the same kind of covariate
data at all.