Corrected construction of path for mesg resource
I've tried to correct problems with trying to get the location of
resource files that supply rif services, especially when the rif
services are being used within other projects.  If you're running a
class in rifServices, then it is easy to navigate within that execution
path to find where files such as RIFServiceMessages.properties is
located.  However, if you're executing something like the desktop RIF
Submission Tool, then even though you're using the service, your
execution path will be within the directory of the rifSubmissionTool
sub-project, not rifServices.  This is an attempt to fix that problem.