Commenting out code for rendering map
The part of the desktop study submission tool that coordinates selected
area IDs in a list with a map was not intended to be developed much.
Instead, emphasis on interactive maps was going to be done in the
web-based version.  However, we should now simply comment out whatever
code there is in the desktop study submission tool that is used to
simulate or fake the idea of an interactive map.  Now, the desktop
submission tool lets you create study and comparison areas by using only
lists of map areas.  They no longer have a visible map component within
the GUI screens.