Band aid fix for the IE JSON problem
We observed that when Chrome was used to call a web service method, the
JSON appeared in the browser.  When IE was used to do the same thing,
rather than display it, the user was prompted to download a JSON
document.  This problem appears to be related to IE's support for
recognising the application/json contentType return in a Response
object.  I've now introduced code which is meant to return a content
type of application/json or text/plain depending on how well we can
detect whether the client is IE or not.  The challenge of doing this in
a good way is well documented, with the main approaches being "browser
sniffing" and "feature detection".  Neither seems to present a great
solution but for now we're just putting in a really basic detection
method that is evident in the class AbstractWebServiceResource, method
generateAppropriateContentTypeResponse.