Adding support for too many areas returned
Modified calls to retrieve map areas so that if there were too many
areas in the results, the code would generate an exception.  The
threshold for what is considered 'too many'' is a property that may be
set in the properties file RIFServiceStartupProperties.properties.
These parameters are used to initialise the value of a threshold limit
that is managed by the RIFServiceStartupOptions object.  This limit is
used within SQLMapDataManager to determine if the number of results
returned exceeds an acceptable limit.  If it does, the code throws an
exception.

Usually I find out how many results there are using a sequence of
resultSet.last() followed by resultSet.getRow().  However, this doesn't
work and when run, this code causes an exception that complains the
result set isn't scrollable.  This suggests there is a property in the
database table which means the result set can only move forward.

For now, we iterate through all the results, count them and if there are
too many, throw a message.  In future it would be far better to find out
if the number of results exceeds the threshold - before we have to
iterate through all of the values.  Perhaps in future it might be better
if we had a version of getMapAreas that only returned a count of how
many results were produced. Alternatively, the result set could be made
to be scrollable to allow the last()...getRow() calls.