Fixing improper exception being thrown
Here, the check for non-existent geo level views and geo level to-maps
had to be fixed.  The order of setting parameters in the
PreparedStatement had to be re-ordered.  As well, when an item was found
to be non-existent, the wrong text and error type were being used to
create an RIFServiceException object.