Made changes to contain rather than extend JTable
Changed classes ending in Table to ensure that they contained rather
than subclassed JTable.  This is an action that is consistent with
containing rather than extending panels and dialogs.  The benefit of
this move is that the behaviour of the class is more clearly defined and
its JavaDocs will not be polluted with references to super class methods
that they don't need.