Changed package name from io to fileFormats
Making a more intelligent name for packages.  IO is a poor choice of
name because that would then include reads and writes to the databases.
The purpose of the fileFormats classes is to read and write to different
file formats.  JSON is not considered so much a format for files but a
format for displaying user interface components.