Adding concept of database properties
RIFServiceStartupOptions is now able to produce a RIFDatabaseProperties
object. (This class is defined in the rifGenericLibrary.dataStorageLayer
package).  The idea is to make this object available to the SQL managers
in rifServices.dataStorageLayer and the query formatter classes in
rifGenericLibrary.dataStorageLayer.  Both of these classes may find it
useful to know whether the RIF is using a PostgreSQL or SQL Server
database.  In fact, this one property of database type could become
important in code switches that use one set of query construction
routines for postgreSQL and another for SQL Server.