Setting up support for identifyDifferences(..)
Eventually most of the RIF business classes will have a method
identifyDifferences(...) that is meant to compare two objects of the
same class - field by field - and produce a collection of messages that
indicate how they differ.  These changes are part of the initial ground
work for that.  The method now appears mostly stubbed in the business
classes.  The class rifServices.util.RIFComparisonUtility has been
developed to support comparisons of two objects.  The properties file
has been enhanced so that it can now support basic messages that explain
the difference between one object and another.