Encapsulated details of createPreparedStatement
Up until now, the SQL manager classes have featured calls such as
connection.prepareStatement(queryText).  However, this call makes
implicit assumptions for the settings of flags related to concurrency
and transaction isolation that may vary from one JDBC driver to the
next.  The code base now isolates the construction of prepared
statements to that flags for resultSetType, resultSetConcurrency and
resultSetHoldability are explicitly set.  This moves allows us to ensure
that regardless of which driver we use, that they will all configure
prepared statements with the same settings.