Fixing library calls related to FormDataParam
Spent significant time identifying problems supporting an upload-file
feature in Jersey.  The problem is that the parameters of the
submitStudy method are not normal String parameters.  Instead, they are
part of a form, which means using the FormDataParam annotation.  I
experienced problems trying to manage the libraries that supported this
but eventually resolved which libraries were needed to support the
feature in Jersey 1.19 libraries.