Added variant of getTiles
We have changed the way we get the database to generate tile data.
Instead of replacing getTiles() with the new implementation, we've added
an additional method that allows us to keep the old method as well.  In
future, getTilesGivenTile may be renamed to getTiles(...) and the old
implementation of the getTiles(..) method will be deleted.