Disabling some validation code
The RIF middleware can serialise and deserialise a RIFStudySubmission
using an XML format.  However, we've realised that when the XML is
coming from the browser client application, it passes fewer data fields
when it submits a study.  For now we're commenting out validation code
for missing fields.  In future we may choose to delete the code, or
develop a "strict" and "relaxed" setting in the checkErrors(...) methods
that appear in the business concept layer.  It might be useful to use a
relaxed policy for receiving a minimum amount of information from a
browser client.  But it might also be useful to retain a strict
validation policy if the middleware were to be sent an XML file that was
produced by some other kind of client that didn't have the same guided
data-entry features.