Making output options non-configurable
Early in development, we thought users may want to choose which
combinations of the following types of output they would want in their
results: data, maps, population holes, ratios and rates.  In the
interest of saving time, we're removing the configurability of these
types.  Every study submission will generate all the different types of
output and users can choose to ignore the results they don't want to
look at.