Removing a class used to test an idea
This is an artefact left over from an investigation about how to
generally improve the performance of middleware methods.  The class was
meant to help cache statements so they could be reused later.  However,
it is much better to rely on using DatabaseProperties and JDBC to
specify caching on the database server rather than trying to maintain a
bespoke one in memory in the middleware.