Specifying strict validation for desktop submission
Validation code developed in the  RIF was created during the time when
we were prototyping the desktop submission tool.  Because the desktop
application was Java-based, it was easy for the client to apply thorough
validation checks to business objects.  When the web-application for the
same tool was developed, it became clear that it would submit far less
data as part of a submission process than the desktop tool.  This is
because the web app tends to submit requests through a few URL
parameters rather than submitting values for every field that could
appear in the data model.  However, we still feel there may be a use to
support two levels of validation.  If the client is capable of
submitting a more complete XML description of the study, the middleware
should oblige and perform as much error checking as it can.  However if,
as in the case of the web application, there are fewer fields, then the
validation policy should be adaptable and perform fewer checks.