Modify middleware call to rif40_startup
Each time the middleware creates a database connection, it must call
rif40_startup(), which is a database procedure that helps create
temporary tables, performs various checks and establishes permissions
for a given user.  Calling this method presents a lot of computational
overhead and currently accounts for most of the time spent in each test
case.  In order to improve efficiency, the database method is now
parameterised so the caller can enable or disable a collection of
checks() which are expensive to do.  Now, when a set of connections is
created for a user, we need only perform these checks once, during the
creation of the first one.  So we no longer have  to perform the checks
every time a connection is created.