Creating a minimal connection manager
Right now, I'm spending a lot of time rapid prototyping the Data Loader
Tool and at the moment I'm not sure how much code the  Data Loader Tool
could or should share with the the rifServices project with respects to
managing connections.  The character of users for rifServices and
rifDataLoaderTool projects seems different.  In rifServices:
(1) there are many users
(2) almost all of the method calls are read-only
(3) sessions are likely short, and time-outs are relevant
(4) a small number of connections is created for each user
(5) a connection is made using an initialisation routine which sets
permissions for a user

In the Data Loader Tool:
(1) There may only be one user
(2) almost all the method calls are write-only
(3) sessions are likely to be very long, especially for long batch jobs
(4) A large number of connections could be made for a single user (the
rif manager) to process large data loads
(5) another initialisation routine may be needed because we need
permissions to create/delete/modify rif tables

I'll know more about how much the SQLConnectionManager used in
rifServices could be used in the data loader tool when more of the data
loader tool is prototyped.  In the meantime, I'm using a very basic one
so I can get on with trying to get the software to execute workflows.