Created test and production loader services
We need a test-version of a data loader service that has methods which
are useful for testing but ones we wouldn't want to include in a
production version.  For example, test cases use the
"clearAllDataSetConfigurations" to reset the database to a known state.
In a production environment we don't include this method!