Adding facility to create/edit data types
Beginning to create the feature for allowing the RIF to support
customised data types.  This was pointed out by the CDC, who wanted to
support data types that may be relevant to their use cases but not
necessarily to Imperial's.  Eg: the UK uses postal codes, the US uses
zip codes.