supporting read-only appearance of text components
Text components can be made editable or non-editable using the
setEditable(...) method.  However, it may not be obvious in a read-only
display that the fields cannot be editable.  Here, we're trying to
centralise the code used to shade the background of text fields a
different colour, in order to indicate that it is read-only.