Enhancing shape file meta data features
Put a lot of thought into whether the DL tool should attempt to open
*.dbf and *.prj to parse important shape file property values or if it
would be better to send the shape file off to the node.js topoJSON
service.  We decided it wasn't worth the overhead of sending traffic
across to a web service when we could just open it on the DL client
instead.  However, we have now begun to rely on the geotools project to
provide a Java class that can parse a *.dbf file that forms part of
shape files.