Fixing validation
Users have the ability to import multiple shape files at once, but they
must remember to properly fill in the properties for each one.  If they
import multiple shape files and then just fill in the currently selected
one, the dialog was allowing this to pass.  Now validation has been
fixed: apart from checking if the currently edited item is correct, it
then checks all the other ones in the list.  If there are any errors (ie
an incomplete shape file that was imported), users are prompted.