Creating generic taxonomy services
This is part of an effort to migrate support for taxonomy services from
the rifServices sub-project to the rifGenericLibrary sub-project.  We're
trying to make the taxonomy service code know as little as possible
about the rest of the RIF, so that it may be used in a more generic
manner.