Initial commit of separated taxonomy services web app
The ability to mark-up study descriptions with health codes is an
important part of study submission.  The activity of fetching terms to
accommodate UI displays (eg: show the root terms, double click on a term
to get subterms etc) is delegated to taxonomy services.  The way these
taxonomy services are implemented is independent of the rest of the RIF.
Therefore, they are being migrated into an independent web service
application in order to (1) better isolate parts of the RIF architecture
that may be reused and (2) enforce a separation of concerns between the
way terms are fetched and the way they are added to study descriptions.