Centralising code for generating nonexistent term
The RIFServiceExceptionFactory is meant to help encapsulate the code
needed to make common exceptions that are thrown throughout the code in
rifGenericLibrary, as well as the other sub-projects that use it.  Here
we're trying to create a generic exception relating to a taxonomy term
that is not known to a taxonomy service.