Putting in synchronised code blocks
Services are initialised not in the constructor but in an initialise
method.  The method should only ever be called once.  To help ensure
that there is no collision for access to the state of isInitialised,
flag, the code used to retrieve or set its state has been put within
synchronised code blocks.