Adding a proxy class to support web service call
The RIF wraps the Java-based middleware APIs with a thin layer of code
that specifically supports web services.  The return value of any method
has to be transformed into JSON, so that it can be rendered in the
web-based front end.  In order to help convert Java-based business
objects into JSON objects, the code includes annotations that can be
used to generate JSON strings.  Including these annotations in the
business classes might bias their design to support a particular form of
deployment.  In order to help them remain agnostic in design, a set of
proxy classes are designed which are simply data containers that are
designed to produce efficient JSON production.  In order to transform a
collection of study summary objects, we have created a proxy class that
can be easily serialised to produce JSON.