Hiding data type declarations in create table
In the interest of promoting better portability between databases, I've
decided to add extra methods to SQLCreateTableQueryFormatter so that the
text phrase used for the actual data type does not have to be specified.
I've then changed whatever current RIF calls make use of the create
table query formatter so they use the new methods.