Changed text type declaration to VARCHAR
The class has a method for adding a text field declaration.  We have a
choice of whether to use VARCHAR or TEXT.  It is recommended in
PostgreSQL to use TEXT for its flexibility and for avoiding VARCHAR(...)
declarations where the size might change.  However, it seems that TEXT
should be avoided in SQL Server.  Therefore, we need to use VARCHAR to
promote better compatibility of using the same query formatter for
different DBs.