Making query formatters schema-aware
Modified AbstractSQLQueryFormatter so that it can now properly consider
the database schema when it is trying to name database tables in the
queries.  For example, instead of rif40_studies, we could now have
rif40.rif40_studies.  When a user creates extract tables for a study,
they would be created in the user's schema - eg: kgarwood.s34_extract