Supporting Integer type checking
Modified the FieldValidationUtility so that it includes methods for
checking if a String parameter is a valid integer.  I developed the
method in response to developing web service calls that rely on a String
value of an integer being passed to it.  Generally, we want all
parameters from a URL to initially be Strings because we want to see if
there had been any attempt to pass malicious code in the values.  After
they have been vetted, the fields are then transformed into parameter
values of the correct type.  In the case of web services that have year
parameter in the URL, first it is checked as a String and then converted
to a numeric year.  The validation method is used to check that the
Integer.valueOf(year) attempt will work.  If it doesn't, then the
FieldValidationUtility will produce a meaningful error message.