Changed to support new data in RIFResultTable
This class is responsible for exracting the data from a RIFResultTable
and serialising it in a JSON string.  Initially, the class only
serialised the two dimensional String array that held table data.  Now,
it can also serialise column names and, using the flag of NUMERIC or
TEXT, it can send back data that can be interpretted as numeric or text
data.