Adding parameter for extra files for extracts
Specifying a start-up property for a directory that will contain extra
files that will be added to each extract file that is generated.  For
example, when a user submits a study and a ZIP file containing the
submission and results is created, that zip file can also be made to
contain a collection of versioned data access documents that were
current at the time when the data set was originally generated.