Centralising code for serialising study results
Created a class that will do two main things: (1) it will write the map
and extract result tables for a study to CSV files.  It then calls on
the RIFZipFileWriter to write the results and the original query, along
with optional extra documents, to a zip file to a directory defined in
the "extraDirectoryForExtractFiles" property of the
RIFServiceStartupProperties.properties file.